Document what sysroot really does.
The previous documentation documented what the original patches for
sysroot did, which was to interpret all file paths as based on the
sysroot path. However, the actual implementation simply removes the
sysroot prefix if a path lexically begins with it. This leads to a
situation where an FcConfig may have a mix of sysroot and non-sysroot
paths. As a result, users should look for a file by first resolving
against the sysroot, but they should also look for the file without the
sysroot.