Add integration with OpenType VF `ital` and `slnt` axes
Map the `FC_SLANT` setting to either the `ital` or `slnt` axis depending
on what is available in the font. This allows client applications with
`FC_FONT_VARIATIONS` support to specify italic or oblique font
variations through fontconfig.

This also introduce a new setting called `FC_SLANT_ANGLE` which is
mapped directly to the value of the `slnt` axis when available. This
allows the user to select which slant angle to be used for oblique
variations.